# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Crypto to Mobile Money Remittance API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Mobile money network
 */
export type CreateOrderRequestNetwork =
  (typeof CreateOrderRequestNetwork)[keyof typeof CreateOrderRequestNetwork];

export const CreateOrderRequestNetwork = {
  MTN: "MTN",
  Airtel: "Airtel",
} as const;

export interface CreateOrderRequest {
  /** Recipient phone number (e.g. 256700000000) */
  phone: string;
  /** Mobile money network */
  network: CreateOrderRequestNetwork;
}

export interface CreateOrderResponse {
  orderId: number;
  /** USDT (TRC-20) wallet address to send funds to */
  address: string;
  message: string;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  waiting: "waiting",
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;

export interface Order {
  id: number;
  phone: string;
  network: string;
  amount?: number | null;
  ugxAmount?: number | null;
  status: OrderStatus;
  txid?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface WalletAddressResponse {
  address: string;
  network: string;
}

export interface ErrorResponse {
  error: string;
}