# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Crypto to Mobile Money Remittance API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: orders
    description: Remittance order operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create remittance order
      description: Creates a new order and returns the wallet address to send USDT to
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateOrderResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/recent:
    get:
      operationId: getRecentOrders
      tags: [orders]
      summary: Get recent orders
      responses:
        "200":
          description: List of recent orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /wallet/address:
    get:
      operationId: getWalletAddress
      tags: [orders]
      summary: Get deposit wallet address
      responses:
        "200":
          description: Wallet address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletAddressResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    CreateOrderRequest:
      type: object
      properties:
        phone:
          type: string
          description: Recipient phone number (e.g. 256700000000)
        network:
          type: string
          enum: [MTN, Airtel]
          description: Mobile money network
      required:
        - phone
        - network
    CreateOrderResponse:
      type: object
      properties:
        orderId:
          type: integer
        address:
          type: string
          description: USDT (TRC-20) wallet address to send funds to
        message:
          type: string
      required:
        - orderId
        - address
        - message
    Order:
      type: object
      properties:
        id:
          type: integer
        phone:
          type: string
        network:
          type: string
        amount:
          type: number
          nullable: true
        ugxAmount:
          type: number
          nullable: true
        status:
          type: string
          enum: [waiting, processing, completed, failed]
        txid:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - phone
        - network
        - status
        - createdAt
        - updatedAt
    WalletAddressResponse:
      type: object
      properties:
        address:
          type: string
        network:
          type: string
      required:
        - address
        - network
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error