# lib/api-zod/src/generated/types/createOrderRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Crypto to Mobile Money Remittance API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderRequestNetwork } from "./createOrderRequestNetwork";

export interface CreateOrderRequest {
  /** Recipient phone number (e.g. 256700000000) */
  phone: string;
  /** Mobile money network */
  network: CreateOrderRequestNetwork;
}