# lib/api-zod/src/generated/types/createOrderResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Crypto to Mobile Money Remittance API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateOrderResponse {
  orderId: number;
  /** USDT (TRC-20) wallet address to send funds to */
  address: string;
  message: string;
}