# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Crypto to Mobile Money Remittance API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  phone: string;
  network: string;
  amount?: number | null;
  ugxAmount?: number | null;
  status: OrderStatus;
  txid?: string | null;
  createdAt: Date;
  updatedAt: Date;
}